Remove token test
There's just no way I can test the received request with the current version of
MockWebServer. That's because the first request received is a websocket upgrade request
with no body. Subsequetly we should listen to messages being sent. That's not
trivial to add in the current implemention. Perhaps it is time for a wholly
different library. Especially since the WS stuff never got merged upstream.